S5PC11X: IRQ: Add ack() call to unmask() for level triggered interrupts.
In Linux, the general sequence for irq handling is:

mask_ack()
irq_handler()
irq_thread_handler() if one exists
unmask()

The irq_handler() or irq_thread_handler() does what is needed to
clear the interrupt.  The problem is that level triggered interrupts
will immediately repend after the mask_ack() before the handlers can
clear the interrupt cause.  We add a ack() to unmask() for level
triggered to handle this case and prevent the second false interrupt
from occurring.

Change-Id: I1e8376300da44fa6bf685ad2b4094693f64dff4a
Signed-off-by: Mike J. Chen <mjchen@sta.samsung.com>