input: cypress-touchkey: driver for touch keypad peripheral
Change-Id: Ib8744466edda655e7e4eec921ec55b98b7710ee7
Signed-off-by: Sang-Jun Park <sj2202.park@samsung.com
Signed-off-by: Jinho Kim <jinho96.kim@samsung.com>
Signed-off-by: Rom Lemarchand <rlemarchand@sta.samsung.com>

General cleanup of the driver, removing unused functions and sysfs interface
since it was agreed upon that userspace wouldn't control the backlight.
Adding delay when turning on the microcontroller to wait for the chip to be
ready.

Signed-off-by: Rom Lemarchand <rlemarchand@sta.samsung.com>

Switching to edge triggered interrupt.
This requires extra code to ignore spurious interrupts while powering on.

Signed-off-by: Rom Lemarchand <rlemarchand@sta.samsung.com>

Cypress made this change to ensure reported key statuses don't get out of
sync between user-space and the controller.
The basic idea is to read the status of all the keys at once on each read and
send all the relevant key states to user-space.

Signed-off-by: Rom Lemarchand <rlemarchand@sta.samsung.com>

Fixing the version check in the touchkey driver to account for older firmware
revs 0xc4 and above.

Signed-off-by: Rom Lemarchand <rlemarchand@sta.samsung.com>