ARM: herring: watchdog: Increase timeout and use realtime workqueue
The old code created a realtime workqueue, but did not use it. This
could cause a reboot if the default workqueue was blocked for more
than one second. Also, the old timeout was short enough to often
trigger a watchdog reset when duming the kernel log in the fiq
debugger.

Change-Id: I556e6bb85dd9a166cab98270a4f1ba23a9f4069c
Signed-off-by: Arve Hjønnevåg <arve@android.com>