S5PC11X: USB: Remove flush the tx fifo on dma
During analysis it has been observed that this is a critical timing issue,
as just during Tethering transfer, UDC driver receives CLEAR_FEATURE request
to clear halt, during this if we put one micro frame delay in setdma_tx path issue
disappears.

Further investigation for the reason behind this delay requirement it has been found that the
last setdma_tx is not getting completed and corresponding complete_tx is not called for it.

Close inspection of setdma_tx it has been found that tx_fifo clearing in setdma_tx function was causing
this issue. So removed fifo flush code from setdma_tx function.

Signed-off-by: sb223.kang <sb223.kang@samsung.com>